// packages/three-modules/src/misc/MorphBlendMeshAnimation.type.ts
export type MorphBlendMeshAnimation = {
  active: boolean;
  currentFrame: number;
  direction: number;
  directionBackwards: boolean;
  duration: number;
  end: number;
  fps: number;
  lastFrame: number;
  length: number;
  mirroredLoop: boolean;
  name: string;
  start: number;
  time: number;
  weight: number;
};